requirements.txt
# ─── Core ─────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.10.6
pydantic-settings==2.5.0

# ─── Database ─────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
psycopg2-binary==2.9.9

# ─── Auth ─────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1              # pinned — passlib breaks with bcrypt>=4.1
python-multipart==0.0.9
google-auth==2.35.0

# ─── OpenAI Agents ────────────────────────────────────────
openai-agents==0.0.7
openai>=1.66.5,<2.0.0

# ─── RAG ──────────────────────────────────────────────────
chromadb==0.5.0

# ─── MongoDB ──────────────────────────────────────────────
motor==3.6.0

# ─── Email ────────────────────────────────────────────────
aiosmtplib==3.0.0

# ─── Scheduler ────────────────────────────────────────────
apscheduler==3.10.4

# ─── Frontend ─────────────────────────────────────────────
streamlit==1.40.0

# ─── Utilities ────────────────────────────────────────────
python-dotenv==1.0.1
httpx==0.27.0

# ─── Testing ──────────────────────────────────────────────
pytest==8.3.0
pytest-asyncio==0.24.0
